fixed some more annotations
Former-commit-id: 2e56ea4fcc31efbed32f0c5f75cc8d59d3754c8a

+++ pp/components/coupler_symmetric.py
@@ -13,7 +13,7 @@ def coupler_symmetric(
     wg_width: float = 0.5,
     layer: Tuple[int, int] = pp.LAYER.WG,
     layers_cladding: List[Tuple[int, int]] = [pp.LAYER.WGCLAD],
-    cladding_offset: int = 3,
+    cladding_offset: float = 3.0,
 ) -> Component:
     """ two coupled waveguides with bends
